重构mycat 2.0 NIO框架以便支持3种线程事务模式(事务与线程绑定,非绑定,reactor),为支持更多事务框架做支持,两种写入模式(透传,自定义报文写入)

+++ mycat2/src/main/java/io/mycat/SQLExecuterWriter.java
@@ -18,8 +18,8 @@ import io.mycat.beans.mycat.JdbcRowBaseIterator;
 import io.mycat.beans.resultset.MycatResponse;
 import io.mycat.beans.resultset.MycatResultSetResponse;
 import io.mycat.beans.resultset.MycatUpdateResponse;
-import io.mycat.datasource.jdbc.resultset.TextResultSetResponse;
 import io.mycat.proxy.session.MycatSession;
+import io.mycat.resultset.TextResultSetResponse;
 import lombok.SneakyThrows;
 import org.jetbrains.annotations.NotNull;
 
@@ -84,7 +84,7 @@ public class SQLExecuterWriter {
                         }
                         case UPDATEOK: {
                             MycatUpdateResponse currentUpdateResponse = (MycatUpdateResponse) resultSet;
-                            int updateCount = currentUpdateResponse.getUpdateCount();
+                            long updateCount = currentUpdateResponse.getUpdateCount();
                             long lastInsertId1 = currentUpdateResponse.getLastInsertId();
                             session.setAffectedRows(updateCount);
                             session.setLastInsertId(lastInsertId1);
